Changes to be committed: 	new file:   LoadImage.py 	modified:   global_vars.py 	new file:   levels_func.py 	modified:   main.py 	renamed:    sprites.py -> player.py 	modified:   scenes.py 	new file:   wall.py
