+++ main.py
@@ -5,55 +5,12 @@ import os
 
 from global_vars import *  # Импортируем нужные переменные
 from scenes import *  # Импортируем сцены
-from sprites import *  # Импортируем спрайты
+from levels_func import *
 
 
 """
 Main loop
 """
-
-
-def load_image(name, colorkey=None):
-    fullname = os.path.join('Data', name)
-    # если файл не существует, то выходим
-    if not os.path.isfile(fullname):
-        print(f"Файл с изображением '{fullname}' не найден")
-        sys.exit()
-    image = pygame.image.load(fullname)
-
-    if colorkey is not None:
-        image = image.convert()
-        if colorkey == -1:
-            colorkey = image.get_at((0, 0))
-        image.set_colorkey(colorkey)
-    else:
-        image = image.convert_alpha()
-
-    return image
-
-
-def load_level(filename):
-    filename = "Data/" + filename
-
-    with open(filename, 'r') as mapFile:
-        level_map = [line.strip() for line in mapFile]
-
-    max_width = max(map(len, level_map))
-
-    return list(map(lambda x: x.ljust(max_width, '.'), level_map))
-
-
-def generate_level(level):
-    new_player, x, y = None, None, None
-
-    for y in range(len(level)):
-        for x in range(len(level[y])):
-            if level[y][x] == '-':
-                Wall(x, y)
-            elif level[y][x] == '@':
-                new_player = Player(x, y)
-
-    return new_player
 
 
 def terminate():
