+++ player.py
@@ -1,11 +1,6 @@
 import pygame
-
 from global_vars import *
-from main import load_image
-
-
-player_group = pygame.sprite.Group()
-wall_group = pygame.sprite.Group()
+from LoadImage import *
 
 
 class Player(pygame.sprite.Sprite):
@@ -128,18 +123,3 @@ class Player(pygame.sprite.Sprite):
             self.image = load_image('./Sprites/stand_sraight.png')
 
         self.change_x = 0
-
-
-class Wall(pygame.sprite.Sprite):
-    def __init__(self, pos_x, pos_y):
-        super().__init__(wall_group)
-
-        self.pf = pygame.Surface((PLATFORM_WIDTH, PLATFORM_HEIGHT))
-
-        self.pf.fill(pygame.Color(PLATFORM_COLOR))
-
-        self.rect = self.pf.get_rect().move(
-            pos_x * PLATFORM_WIDTH, pos_y * PLATFORM_HEIGHT)
-
-    def update(self, screen):
-        screen.blit(self.pf, self.rect)
